refact: useDynamicModalSize 훅으로 파일 분리
